Use `indent` procedure of `ElementTree`
Pretty printing has been implemented with the `toprettyxml` function
that has unintended side effects with texts. From Python 3.9 the
`intend` procedure is available on `ElementTree`. If this is available
and Python 3.9 or higher is used, the code uses this procedure.
Otherwise, a fall back to the old code is used.